requirements.txt
python-telegram-bot==20.7
pdf2docx==0.5.6
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.1.0
pymupdf==1.23.8
python-dotenv==1.0.0
